+++ lib/src/ioc_container.dart
@@ -1,4 +1,16 @@
+import 'dart:async';
+
 import 'package:get_it/get_it.dart';
+
+/// Describes the type of change that occurred which triggered the
+/// [IocContainer.OnScopeChanged] callback.
+enum ScopeChange {
+  /// Indicates a new scope was created and is now active.
+  created,
+
+  /// Indicates a scope was removed and its parent scope is now active.
+  removed,
+}
 
 /// A standard interface providing inversion of control services to Dart or
 /// Flutter applications.
@@ -19,6 +31,23 @@ abstract class IocContainer {
   /// Allows registering a custom implementation of the [IocContainer] interface.
   static void registerContainer(IocContainer iocContainer) =>
       _container = iocContainer;
+
+  /// Gets the name of the current scope.
+  ///
+  /// If the current scope doesn't have a name `null` if returned. If the
+  /// current scope is the root scope 'root' is returned.
+  String? get currentScope;
+
+  /// The callback that is called when the current scope has changed.
+  ///
+  /// Detecting scope changes can be helpful in situations where depending code
+  /// needs to updates its dependencies based on the registration in the new
+  /// scope.
+  ///
+  /// The [scopeChange] parameter supplied to the callback indicates if a new
+  /// scope was added (see [createScope]) or an existing scope was removed (
+  /// see [removeScope]).
+  void Function(ScopeChange scopeChange)? onScopeChanged;
 
   /// Gets an instance matching the specified type [T].
   ///
@@ -104,7 +133,7 @@ abstract class IocContainer {
   /// a bug. However in some special cases this might be valuable and allows
   /// developers to override an earlier registered factory.
   void registerFactory<T extends Object>(
-    FactoryFunc<T> func, {
+    T Function() factoryFunction, {
     String? instanceName,
     bool allowReassignment = false,
   });
@@ -165,10 +194,15 @@ abstract class IocContainer {
   /// value is `false` as it would normally not be required and in most cases is
   /// a bug. However in some special cases this might be valuable and allows
   /// developers to override an earlier registered singleton.
+  ///
+  /// The [onDispose] callback is called when the lazy singleton is destroyed
+  /// (for example when the [reset], [resetLazySingleton] or [remove] methods
+  /// are called) and can be used to clean up additional resources.
   void registerLazySingleton<T extends Object>(
-    FactoryFunc<T> func, {
+    T Function() factoryFunction, {
     String? instanceName,
     bool allowReassignment = false,
+    FutureOr<void> Function(T)? onDispose,
   });
 
   /// Registers a specific instance of type [T].
@@ -221,11 +255,93 @@ abstract class IocContainer {
   /// value is `false` as it would normally not be required and in most cases is
   /// a bug. However in some special cases this might be valuable and allows
   /// developers to override an earlier registered singleton.
+  ///
+  /// The [onDispose] callback is called when the singleton is destroyed
+  /// (for example when the [reset] or [remove] methods are called) and
+  /// can be used to clean up additional resources.
   void registerSingleton<T extends Object>(
     T instance, {
     String? instanceName,
     bool allowReassignment = false,
+    FutureOr<void> Function(T)? onDispose,
   });
+
+  /// Resets an earlier created instance of a registered lazy singleton.
+  ///
+  /// Resetting the singleton will cause the IoC container to call the
+  /// registered factory function again the next time the [get] method is
+  /// invoked. This will recreate a new singleton.
+  ///
+  /// Selecting the lazy singleton to reset can be done by providing the
+  /// [instance] to reset, providing the registered type [T] or the
+  /// [instanceName] the lazy singleton was registered with.
+  ///
+  /// ```dart
+  /// class Counter {
+  ///   Counter({this.count = 0});
+  ///
+  ///   int count;
+  ///
+  ///   void increment() => count++;
+  ///   void decrement() => count--;
+  /// }
+  ///
+  /// // Register a factory method to create a new instance of the Counter
+  /// // class starting at 0.
+  /// IocContainer.container.registerLazySingleton<Counter>(() => Counter(), instanceName: 'zero_based_counter');
+  ///
+  /// // Get a new zero-based Counter instance from the IocContainer.
+  /// Counter counter = IocContainer.container.get<Counter>(instanceName: 'zero_based_counter');
+  ///
+  /// // Increase the current count.
+  /// counter.increment();
+  /// // The current count should by increased to 1.
+  /// print('Current count: ${counter.count}');
+  ///
+  /// // Reset the lazy singleton
+  /// IocContainer.container.resetLazySingleton<Counter>();
+  ///
+  /// // Get the Counter instance from the IocContainer.
+  /// counter = IocContainer.container.get<Counter>(instanceName: 'zero_based_counter');
+  ///
+  /// // Increatee the current count.
+  /// counter.increment();
+  /// // The current count should be incremented to 1 as a new instance was created.
+  /// print('Current count: ${counter.count}');
+  /// ```
+  ///
+  /// If resources need to be disposed before the reset, provide a
+  /// [onDispose] callback. This function overrides the disposing
+  /// you might have provided when registering.
+  FutureOr<void> resetLazySingleton<T extends Object>({
+    Object? instance,
+    String? instanceName,
+    FutureOr<void> Function(T)? onDispose,
+  });
+
+  /// Creates a new registation scope.
+  ///
+  /// Registering types after creating a new scope they will hide any previous
+  /// registrations of the same type. Scopes will allow for managing different
+  /// live times of your Objecs. Scopes are stacked upon eachother, when an
+  /// Object is retrieved it will look for it in the top most scope (the one
+  /// created last) first, if the Object insn't located it will look into the
+  /// next scope. This will continue untill the Object is found or until there
+  /// are no more scopes (which will result in an error).
+  void createScope({
+    String? scopeName,
+    FutureOr<void> Function()? onDispose,
+  });
+
+  /// Removes the current scope from the [IocContainer].
+  ///
+  /// Removing the current scope will dispose all factories and singletons
+  /// registered within the scope and makes the parent scope active again.
+  ///
+  /// Provided dispose callbacks on factories and singletons will be called
+  /// when removing the scope. If a dispose callback was supplied when
+  /// creating the scope it will also be called.
+  Future<void> removeScope();
 }
 
 /// An implementation of the [IocContainer] interface using get_it as
@@ -236,6 +352,11 @@ class _GetItIocContainer implements IocContainer {
   _GetItIocContainer._();
 
   final GetIt _container = GetIt.instance;
+
+  @override
+  String? get currentScope => _container.currentScopeName == 'baseScope'
+      ? 'root'
+      : _container.currentScopeName;
 
   @override
   T get<T extends Object>({
@@ -259,10 +380,14 @@ class _GetItIocContainer implements IocContainer {
     FactoryFunc<T> func, {
     String? instanceName,
     bool allowReassignment = false,
+    FutureOr<void> Function(T)? onDispose,
   }) =>
       _guardedReassignment(
-        () => _container.registerLazySingleton<T>(func,
-            instanceName: instanceName),
+        () => _container.registerLazySingleton<T>(
+          func,
+          instanceName: instanceName,
+          dispose: onDispose,
+        ),
         allowReassignment,
       );
 
@@ -271,12 +396,41 @@ class _GetItIocContainer implements IocContainer {
     T instance, {
     String? instanceName,
     bool allowReassignment = false,
+    FutureOr<void> Function(T)? onDispose,
   }) =>
       _guardedReassignment(
-        () => _container.registerSingleton<T>(instance,
-            instanceName: instanceName),
+        () => _container.registerSingleton<T>(
+          instance,
+          instanceName: instanceName,
+          dispose: onDispose,
+        ),
         allowReassignment,
       );
+
+  @override
+  FutureOr<void> resetLazySingleton<T extends Object>({
+    Object? instance,
+    String? instanceName,
+    FutureOr<void> Function(T)? onDispose,
+  }) =>
+      _container.resetLazySingleton(
+        instance: instance,
+        instanceName: instanceName,
+        disposingFunction: onDispose,
+      );
+
+  @override
+  void createScope({
+    String? scopeName,
+    FutureOr<void> Function()? onDispose,
+  }) =>
+      _container.pushNewScope(
+        scopeName: scopeName,
+        dispose: onDispose,
+      );
+
+  @override
+  Future<void> removeScope() => _container.popScope();
 
   void _guardedReassignment(
     void Function() register,
@@ -289,5 +443,25 @@ class _GetItIocContainer implements IocContainer {
     register();
 
     _container.allowReassignment = false;
+  }
+  
+  void Function(ScopeChange scopeChange)? _onScopeChanged;
+
+  @override
+  void Function(ScopeChange scopeChange)? get onScopeChanged => _onScopeChanged;
+
+  @override
+  set onScopeChanged(void Function(ScopeChange scopeChange)? onScopeChangedCallback) {
+    _onScopeChanged = onScopeChangedCallback;
+
+    _container.onScopeChanged = (bool pushed) {
+      void Function(ScopeChange scopeChange)? callback = _onScopeChanged;
+
+      if (callback == null) {
+        return;
+      }
+
+      callback(pushed ? ScopeChange.created : ScopeChange.removed);
+    };
   }
 }